fix: adapt tests to use SSE server instead of stdio
Also, do not error out if we fail to create a second global subscriber -
in the mcp server it appears to get to that stage a second time. Ideally
we'd find out how and stop that, but for now it isn't really that
important so better to let the error get swallowed imo.