chore: add pre-project suggestion to LLM
When testing the MCP integration, I found it generated a pyproject.toml
file with `hatchling`, which is both not needed, and broke the project.
Since we really only need the basics for a project, `uv init` is just
fine.

We could also add a tool for this in the MCP theoretically, since we
install uv for the user and run it behind the scenes, but this is the
easier change for now (and as long as they have uv installed, which I
would guess most Python peeps these days would, it would work fine).