fix: remove timeout test
We removed the timeout in 8f8cb29186a6e6ac59e68ce36e003c5c7eb0cd23
(rebased away by now). The reason was that it was a timeout for the
entire app run, at which point it was v. inappropriate. Even CI systems
only do it on _output_ inactivity, not total inactivity. Secondly it was
the wrong location for implementing a timeout, which is clearly
something on the MCP or CLI level of concern, not the lower down the
stack monitor_output().

But the thing I don't get is... here was the test the entire time,
running fine in fix/cli-run-regressions. But there was no timeout
mechanism. WTF?