fix: make CLI less brittle to erroneous nulls
When we see a null value for a required field, the generated code will
now instruct serde to give it the default value (e.g. "" for string, []
for array, etc). This may be nonsense, of course, but it's less likely
to be nonsense than "null", and makes the CLI a whole lot less brittle
to bugs on the backend - a single null value shouldn't kill the rest of
a response!